made alot of updates
added a bbunch of jquerry to wrap and collapse a more info button, as well as hide the standard read more div. Also styled the button. Cannot seem to rotate the arrow down so I'm calling it a night. This will easily translate to a video and venues button. Next is to build out the venues feature, which will take time.